user_agent/requirements.txt
openai>=1.30.0
langchain-mcp-adapters>=0.0.6
pillow>=10.3.0
redis>=5.0.7
